+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## Unreleased
-### Fixed
-- Fixed bug with server-side debugger when game was in deferred events mode
+## [0.6.1] - 2022-07-10
+### Added
 - Added `Matter.log`, and a logs viewer per-system to the Matter debugger.
 - Added error logging and inspection to the Matter debugger.
 - Added Query inspection to the Matter debugger.
+
+### Fixed
+- Fixed bug with server-side debugger when game was in deferred events mode
 
 ## [0.6.0] - 2022-07-08
 ### Changed